refactor and fix terminal data types
