+++ src/createMollieClient.ts
@@ -203,5 +203,5 @@ export { RefundEmbed, RefundStatus } from './data/refunds/data';
 export { SubscriptionStatus } from './data/subscriptions/data';
 export { ProfileStatus } from './data/profiles/data';
 export { OnboardingStatus } from './data/onboarding/data';
-export { TerminalStatus } from './data/terminals/Terminal';
+export { TerminalStatus } from './data/terminals/data';
 export { default as MollieApiError } from './errors/ApiError';
